Update with requirements of Coding_Exercise_3
In exercise 1 you created a function that converted Celsius degrees to Fahrenheit. 

The lowest possible temperature that physical matter can reach is -273.15 °C. With that in mind, please improve the function by making it print out a message in case a number lower than -273.15 is passed as input when calling the function.